Audit-AzureWebAppMinimumTlsVersion
This script takes either subscription name(s) as commma seperated values and audit all azure web apps in those subscriptions. 
Output will be exported to a csv file in the execution folder of the script.
If no subscription names are provided as input, script will get list of subscriptions and execute audit on all subscriptions.